(PDB-697) Document and test new events and resources features
This change adds tests and docs for events/resources features inherited
by the switch over to the query engine.

This change also documents and tests an explicit "null?" operator. Some
endpoints for v2/v3, when using not, such as ["not" ["=" "line" 10]]
would return line values that were NULL in addition to other line
numbers. In v4, this is no longer the behavior. Users wanting null
values in addition to lines != 10, can use "null?" with an "or"
statement.