(maint) Restrict which process will be stopped
This is kind of an after-thought, but we figured since this file is
already being edited, this would be a good opportunity to add in this
additional failsafe. The idea is that no puppetdb process should be
started by anyone except the puppetdb user. If, for some reason, the
puppetdb process fails, or is stopped, but its PIDfile isn't removed,
there is the possibility that some other process can take its old PID.
In that case, stopping the puppetdb service would look at the PIDfile,
and stop the wrong process because it had taken on the old puppetdb PID.
Granted, this situation has a very small chance of occuring, especially
since we also limit what process will be stopped with the `--exec` flag.
However, it's nice to safeguard against the possibility anyway.