(docs) Normalize multi-paragraph lists in release notes
Markdown supports multi-paragraph list items, where paragraphs after the first
are indented. According to the original spec (such as there is), the indent must
be four spaces, but most renderers will be flexible and accept two.

Unfortunately, other renderers are sticklers and demand their four spaces, and
we're considering moving to one of those (because it's faster than the others and
lets you do GitHub-style code blocks). Thus, this commit switches to four-space
tabs on big list items.