(PDB-708) Add backwards compatible querying capability for facts
This patch fixes our queries so that the old facts end-point still returns
facts.

It uses the new schema, but is only able to display items at the top level
for now. It does this by checking a depth column we populate upon command
submission to fact_paths.

As a consequence of this change, it creates some question about backwards
compatibility. For now I'm reverting the change to unstringify the facts
at the terminus so the old behaviour is still kept end-to-end.

Signed-off-by: Ken Barber <ken@bob.sh>