(PDB-550) Update puppetDB docs to include info on LibPQFactory
Prior to this commit, the docs only enabled ssl for postgresql
connections. This commit changes the docs to add a few extra lines in
the database.ini file to use host cert/key and CA cert for verification
with ssl.

This commit also shows how PDB-550 can be accomplished, since it loads
PEM files and uses SSL without the use of keystores (yay!).