(maint) Switch confine for basic test during acc dependency installation
The way we were using confine was wrong, and since this is now more strict
in beaker it was throwing errors in the master smoke tests. This patch
just replaces it for a basic include? on the master platform instead.

Signed-off-by: Ken Barber <ken@bob.sh>
(cherry picked from commit 42f6ddfd1934b887984c381fc080fa883e83abc2)