(PDB-826) Fix pathing for puppetdb-legacy
For PE the puppetdb-legacy scripts (such as puppetdb-ssl-setup) expect the
puppetdb script to be in the path. This assumption isn't always true, as we
install PE in a weird place.

This patch adjusts the PATH for this single exec so that it also includes
the path of the directory where the original script is found, which is usually
/opt/puppet/sbin.(PDB-826) Fix pathing for puppetdb-legacy

Signed-off-by: Ken Barber <ken@bob.sh>