(maint) Fix regexp array operator test failures for PostgreSQL
PostgreSQL deals with lookaheads differently to HSQLDB, this work-around
switches a .* to an or'd regexp that matches in functionality but still
retains the expected look ahead behaviour. This is all to ensure we still
pin regexp lookups to within the delimiters.

I've also fixed that case where users might provide a number and expect
that to be matched exactly.

Signed-off-by: Ken Barber <ken@bob.sh>