(maint) Update rpm spec and systemd service file for pidfile handling
Previously, the rpm package was creating /var/run/puppetdb at install time for
pidfile storage. This will break on modern Fedora/RHEL and SUSE systems as
/var/run (which is a symlink to /run) uses tmpfs and will not persist through a
reboot.

As the systemd service type is "simple", meaning non-forking, systemd
manages the process directly, and the PIDfile is not needed or used on
those platforms.