(PDB-813) Upgrade trapperkeeper, and move certificate-whitelist
This patch upgrades trapperkeeper, and the jetty component to the latest
revision. All lines of this patch are attributable to getting this to
work.

The major change was the change to certificate-whitelist, which is
not supported in trapperkeeper-webserver-jetty9 so we had to intercept
the configuration item, deprecate it and move it to a new [puppetdb] section
so we could do this work ourselves.

We avoid using [global] in this instance, since we have identified
a potential to clash if we want to bundle PDB with other clojure applications,
this is a general problem anyway - but our first move to try and reduce
this potential.

Signed-off-by: Ken Barber <ken@bob.sh>