(PDB-882) Setup basic facts performance testing
This introduces our first pass at performance testing, in particular we have
targetted loading structured facts.

The benchmark tool is now capable of using a series of provided fact JSON files,
so we can use that tool for this purpose. We've also provided a set of basic
seed files for these tests.

The tooling: `rake test:beaker` has now become `rake beaker:acceptance` and
performance tests are started using `rake beaker:performance` instead.

The testing introduces a new host configuration specific to this work, in particular
we are using a single c3.xlarge instance. This is isolated from the acceptance testing
one so we can tune this on its own.

This also fixs a number of side issues:

* benchmark -h was broken, now fixed
* The -rp switch no longer worked, this has no become -r
* Some of the client based stuff was strewn around our code base, its now
  centralized in puppetlabs.puppetdb.client

Signed-off-by: Ken Barber <ken@bob.sh>