Try reverting to c1.medium
After the change to c3.large we started to see intermitent issues with tests,
this patch is to just test the theory that it might be the box change
somehow changing the timing characteristics or some such. We're not 100%
so we might need to revert this also.

Signed-off-by: Ken Barber <ken@bob.sh>
(cherry picked from commit 3aa02cfb5c132b968cc20610061bb65d53411d62)