More clarification around using an IP address for ssl-host and host
Setting `ssl-host` and `host` to a hostname often causes problems due to
resolution failing, pointing at the wrong thing or sometimes even resolving
first to an IPv6 address instead of an IPv4.

This patch fixes some documentation and the comment in the configuration
file to forewarn users of this scenario, so they can avoid it.

Signed-off-by: Ken Barber <ken@bob.sh>