Correct the ssl-protocols rewrite for the master branch
Right now tk-webserver-jetty9 0.7.5 requires a list to be passed to it for
the setting `ssl-protocols` but our rewrite still provides a comma separated
string.

While this is a bug in-and-of-itself, its causing our tests to fail. This
patch gets us back to passing. We'll look at the configuration issue as a
part of the fixes going into TK that might resolve this.

Signed-off-by: Ken Barber <ken@bob.sh>