(PDB-168) Improve error messages from .ini file parsing.
- Split hook-tk-parse-config-data and put the bulk into hook-tk-parse-config-data'
so I could pass in an action-on-error-fn.  In production this will be a System/exit
call, but test code uses it as a sort of callback mechanism to validate exception
handling.
- Added code to handle 3 potential exceptions thrown by TK parse-config-data.  The
full exception is logged, a clean+simple error is reported to the console, and the
production system will exit immediately.
- Added test code to induce each exception condition, and verify expected handling.
- Validated expected production and test behavior with missing/empty/invalid
configuration files, using 'lein run services -c config.ini' and
'lein test puppetlabs.puppetdb.config-test' commands.