(PDB-1038) Include a new namespace for ezbake to use for service startup
This provides a new namespace: puppetlabs.puppetdb.main that will be used to
startup the PuppetDB service. This avoids the need to include the sub-command
'services' in the ezbake configuration templates, avoiding the need for
a new parameter to be added to ezbake to support this 1 edge case.

The new namespace contains a single function -main, that effectively wraps
the sub-command 'services'.

We also noticed that the core.clj file had a gen-class declaration, since
we don't use AOT this is not needed and is being removed.

Signed-off-by: Ken Barber <ken@bob.sh>