(PDB-1034) Ezbake source based testing
This change switched PuppetDB over to use the ezbake_utils helpers in ezbake
to install PuppetDB via source. This effectively means the old installation
methodology can be retired.

For now, this just pins us to a branch for beaker, until that patch is merged.

The beaker-tests-source.sh script is now modified to ensure that leiningen
is now installed (since its required by ezbake helpers in beaker).

Some minor code was removed that is no longer needed, but a major retirement
patch will be forthcoming to remove the greater amount of code no longer
required. Since it was deemed such a patch would make this patch hard to read.

Also, retry was added to the acceptance test jobs, to ensure we can avoid
transient issues getting at the git version of beaker.

Signed-off-by: Ken Barber <ken@bob.sh>