Ensure we use our own vendored directory for gems
We were using the global gemset which is a bad thing, this patch fixes that.

This was causing issues when a fog-core gem that was bad had been permanently
'cached' due to the use of a global gemset. This will slow the build down
and create another potential place of failure, but it is more correct.

Signed-off-by: Ken Barber <ken@bob.sh>