(PDB-1034) Switch to using a script in the repo for PR testing
This switches us to using a script sources from the PuppetDB repo for PR
testing instead of storing the script in the jenkins job. This allows us
to easily do very different things if the branch is different, and will
prepare us for the PDB-1034 work where the actions required to bootstrap
testing for stable/master will be slightly different.

Signed-off-by: Ken Barber <ken@bob.sh>