(PDB-775) Tidy up acceptance helper for package based installs
This commit fixes an issue with package based acceptance tests
whereby we weren't applying the correct manifest for managing puppetdb
on either embedded or postgres puppetdb installs. The hsql issue was due
to the fact we weren't passing in `:embedded` to the module causing the
module to try and validate a postgres database, the postgres issue was
due to the fact that the db param to install_puppetdb is a keyword and
not a string so we were always applying the incorrect hsqldb manifest.