(PDB-840) Retire deprecated commands
This commit retires all of the previously deprecated commands and
removes all the associated support code for these commands, such as
validation of reports/catalogs for the deprecated versions. This commit
also updatees al the unit/integration tests to use the non-retired
versions of these commands and removes test code we no longer use.