(PDB-1072) Add endpoint to mirror report command structure.
This commit changes the /v4/reports endpoint to mirror the store-report command
structure -- this amounts to an additional field containing resource-events.
The upshot of this is that the api response can be easily munged for resubmission
as a report. In an HA context, this will allow us to transfer reports from one
database to another over HTTP without needing a second query for the events
associated with a report.