(maint) Allow factset ordering on hash
This corrects an oversight in adding the hash field to factsets that left
ordering results on hash impossible. I'm not a fan of the solution here, but it
follows our approach for environments and I can't think of a better way ATM.
The constraints are:

a) The factsets query must be "doubled up" so that paging options are applied
on an object basis instead of a by-row basis.

b) Both environment and hash are nullable fields in factsets, so a COALESCE is
necessary to prevent a "WHERE null IN null" scenario.