Apply the TK hook config to the test instances too
Previously the jetty testutils service instances' configs were only
adjusted by process-config!, but the actual server's config was also
adjusted by hook-tk-parse-config-data.  Change it so that
adjust-tk-config is applied in both cases.

Doing so exposes a bug -- a [global] url-prefix setting with a relative
path like "foo" will prevent the server from responding on /foo.  That's
because add-web-routing-config (via the tk hook) wasn't normalizing the
url-prefix before stashing it in in the web-router-service config.  To
fix that, normalize the url-prefix in add-web-routing-config.