(PDB-1070) Deprecate url-prefix in global
This commit deprecates the use of the `url-prefix` config setting under
the `[global]` header, favoring the use of the webrouting service to
configure the url-prefix. This commit uses the WebroutingService
get-route function to retrieve the url-prefix from the config data,
using the hook in config.clj to add a default or copy the config from
global url-prefix to the web-routing config.