(PDB-800) Move metrics server to its own tk service
This commit moves the metrics endpoints to their own trapperkeeper
service, using the web-router-service to mount it and the
puppetdb service at the appropriate routes. This commit also addresses
PDB-200 by combining the mbean/mbeans endpoints to a single mbeans
endpoint.

This commit moves the metrics documentation to its own top-level
heirarchy on docs.puppetlabs.com to reflect the fact that the metrics
endpoints have been moved to their own namespace.