(PDB-908) Improve debugging info of in-process launched PDB
This patch dynamically creates a temp logback and temp log file that an
in-process launched PuppetDB instances uses. Previously when an error
occurred in that Jetty instance, it was difficult to determine the
problem. Sometimes errors were lost, others were difficult to sort
through since it wrote to the same System.err that the test thread was
running in. This patch writes all in-process PuppetDB log information to
an actually file, which is scraped and dumped into the test thread's
console, make it much easier to diagnose errors