(PDB-1109) Couldn't satisfy such that unit test failure
The couldn't satisfy such-that unit test failure can happen if the
test.check generator generates 10 possible "trees" that are not
collection, such as all leaves. This can cause false negative test
results. This commit increases the probability a tree with leaves will
be created and increases the number of attempts (from 10 to 100) that
such-that will make before giving up.