(PDB-1067) Warn user with unknown config items
This commit changes the behavior of the config validation by warning
users in the log when there is an unused config item in the config file
instead of failing hard. This commit also refactors the items in
config.clj so that changes like this are a little easier in the future
by adding a `configure-section` function which removed the duplication
with `configure-puppetdb`, `configure-write-db`, etc.