(patch) Turn off strict_validation for master::config
This commit addresses an issue where v3 of the http api was removed from
pdb but a bug in the pdb-module prevents us from validating a connection
from the master using the v4 of the api without a release of the module.
This *should* be turned back to the default (true) when this issue is
addressed. Also this commit removes the $test_url param from being used
since it misleadingly never gets used by the provider in the module.