(PDB-332) Add tags to catalog hash
If the user submits a new catalog with no changes other than tags, 
the catalog should still be stored. It previously was not, since the
tags were not part of the catalog hash computation. 

This does not update existing hashes stored in the db, so the next 
catalog update for each node will cause a db write.