(PDB-908) Adding a new test until-consumed function
These changes are to the tests only. This function will invoke a
function and wait for the messages put on the queue to be consumed. Will
throw an exception if the message takes too long, ends up in the
DLO. There is also code to ensure that the JMX beans have been
initialized after startup before the checking for consumed messages.