(PDB-1241) Resource Event timestamps should be normalized before stored
There are many ways to represent an ISO 8601 timestamp in string
form. The end result is the posibility of having to two timestamps that
are the same, but represented different which causes the report to hash
differently. Although Puppet should be sending PuppetDB UTC dates, our
tests were using non-UTC dates, which would cause the dates to be stored
in another timezone, then converted to UTC on the way out of the
database which is really confusing.