(PDB-1259) Faster conversion of strings to timestamps
Previously we were using clj-time.format/to-timestamps's default
ordering of date formatters to attempt to parse the incoming
string. This just used the values in the formatters map in whatever
order they are stored. These formats are used in the order they come out
of the hash-map until one successfully parses the time. This means we
could potentially have 10+ failures to parse a date before a successful
parse, and we do this each time in each report/catalog/factset etc.

This commit switches to use an ordered list of formatters, using the
most likely parsers to work (i.e. ISO8601), then falls through to the
existing behavior.