(PDB-1255) Switch tests to pick open port before bootstrapping TK
Previously tests would set Jetty's port to 0, which automatically picks
an open port on startup. The downside of that was the PuppetDB config
wouldn't be updated with the actual port chosen, so that info would have
to be pulled from the Jetty server. This patch switches to determine an
open port before starting TK, then use that port in the TK config. This
makes our test more like what users would run (specifying a port rather
than letting Jetty choose).