(PDB-1179) Simplify start-puppetdb and webapp building
This commit cleans up the syntax of the starting of puppetdb and the
building of puppetdb webapplications with an authorizer. As new
webservers are added into the puppetdb project it is desirable to have
clean, standardized examples of build-app functions and configuration
manipulation. This commit also adds the authorizer, built from the
certificate-whitelist config, to the shared-globals of the
PuppetDBServer so that other apps need not build the authorizer from
scratch again.