(PDB-1286) backport resource timestamp normalization to stable
In master, we've changed our report hashing so that timestamps are converted to
a consistent representation before being hashed.

This backports that change to stable so that users will receive the update in 2.3.1,
and by the time they make it to 3.0 their reports should already be right.

Additional handling will be required in master for various edge cases.