(PDB-1342) Fix trusted facts fallback
When adding trusted facts, a fallback mechanism is used to pull them
from the node object if there aren’t any there already. This code is
broken in puppet 4; it’s unclear whether it happened to work in puppet 3
because there was a global request object available or because this 
code was never called. In any case, the change works in both.