Update acceptance tests to remove global manifest
As of Puppet 4, the use of `manifest` as a global var is no longer
supported.

This commit updates the acceptance helper and tests to build out
dynamic environment paths and set an `environmentpath` var as a
replacement to the previous technique. It now uses the
create_remote_site_pp helper function prior to using
with_puppet_running on. This function ensures that an environment
path is created on the SUT and returns this path to be used in the
manifest.