(PDB-1308) Change to pull-based sync
Change the sync protocol to be entirely pull-based. Previously, the source host
would initiate sync by sending a summary record to the destination host. Now
the destination host uses the query API to retreive the same information. This
both simplifies the sync protocol and makes it easier to rebuild a PDB when it
joins a cluster after losing its data.