(PDB-1300) Get tests working with Puppet 4 AIO packaging
This patch adds support for acceptance tests to run against the Puppet 4 AIO
packaging.

It includes conditionals for BEAKER_TYPE=aio so that these divergent setup/pre-suite
tasks can occur, and includes a special puppetserver_embedded_db option for
embedded DB support only (hsqldb), with the goal to follow up with postgresql
support soon.

The core of this change is around how we install puppet, and manage puppet runs.
This code provides conditional facilities around existing functions to make sure
this is transparent, or at least minimally opaque to the tests.

All tests that needed modifications have been modified in a minor way to work
with both this new setup, and the old Puppet 3 setup so there is no
compatibility issues and we can continue to test on both.

We also fix a minor bug, whereby tests that included seperate database and master
nodes would fail due to the puppet node deactivate command being executed on
the wrong host. This wasn't being surfaced in our pipeline since we don't test
this case today, but it was found and fixed regardless.