(PDB-1389) Add puppet-agent option for debian dependencies
Previously installing the puppet-agent and puppetdb together would not
work on debian/ubuntu because puppetdb requires a specific version of
puppet or puppet-common, which puppet-agent cannot provide. This commit
addresses that by allowing puppet-agent to be a fallback if
puppet/puppet-common is not available (puppet for puppetdb,
puppet-common for the terminus).