(PDB-1416) Change the postgres type for hashes and uuids
UUIDs stored as strings are about 36 bytes, and are 16 bytes stored as
the native type. Hashes are 40 bytes as characters, but only 20 has a
bytea field.
This commit changes PuppetDB to store the uuid's as their native type
and hashes as the bytea type (we can easily interact with the bytea type
by using the hexidecimal shorthand, prefixing hashes with '\x'.
This commit also enforces that our uuids be non-nil as all supported
versions of puppet enforce this.