(PDB-1428) Suppress version-check INFO logging
The version-check function uses clj-http, which uses Apache HttpClient
for all the heavy lifting. Recent versions of HttpClient log connection
problems to INFO when using the library's built-in retry facilities.
This is suboptimal, as what's supposed to be a silent background process
(one where it's okay if it fails) to something that may freak out users
with scary looking error messages that are actually harmless.

If you disable automatic retries on the request, HttpClient throws an
exception; this matches the behavior of older versions. Since retries
aren't really required for version-checking, this patch disables them
and thus returns control of logging back to us.