Explicitly add hostname and fqdn to dns_alt_names
Beaker is not including the hostname of the SUT in the dns_alt_names
on the puppet master. This is causing the cert lookup of the master
node to fail in ec2.

This commit explicitly adds the master's hostname and fqdn to the
dns_alt_names to make sure that the cert can be found by these
names.