(maint) Speed up storage.clj `<hash> in (?,?,?...)` queries
This commit addresses an issue where the above query for both
`resources` and `facts` was very slow on Postgres because we were
trimming the bytea hash to a string which was causing Postgres to do a
table scan. This commit fixes this issue by changing the parameters in
the `(?,?...)` to be munged bytea hashes so we don't do a table scan.