(PDB-1447) remove late project
This commit adds a migration to change the value_json column of facts to a TEXT
"value" column, which stores a json-serialized copy of the properly typed fact.
This field is returned by facts queries and deserialized in code to avoid
dragging type information out of the database. Doing this allows us to remove
the concept of late-project.