(PDB-1448) GC values that swap paths correctly
Make sure that the values GC in replace-facts also considers value
references that are established by any swapping of paths in the current
call.

For example, in cases where replace-facts was replacing these
facts (factset path value):

  (0 "x" 1)
  (0 "y" 2)

with these:

  (0 "x" 2)
  (0 "y" 1)

and there were no other paths referring to to the values 1 and 2 in the
database, the replace-facts values GC would fail with a postgresql
constraint violation similar to this:

  ERROR:  update or delete on table "fact_values" violates foreign key constraint "fact_value_id_fk" on table "facts"
  DETAIL:  Key (id)=(1) is still referenced from table "facts".
  STATEMENT:  DELETE FROM fact_values fv
                     WHERE fv.id in ($1,$2)
                       AND NOT EXISTS (SELECT 1 FROM facts f
                                        WHERE f.fact_value_id in ($3,$4)
                                          AND f.fact_value_id = fv.id
                                          AND (f.factset_id, f.fact_path_id) NOT in (($5,$6),($7,$8)))

because the final AND clause excludes all paths referring to 1 and 2
from consideration as live references, when it should only be excluding
the specific references that are being deleted.  In this example the
exclusions should be (0 "x" 1) and (0 "y" 2), instead of just (0 "x")
and (0 "y").

Include the value in the final clause, i.e. (factset_id, fact_path_id,
fact_value_id), to make the exclusion appropriately specific and fix the
problem.

cf. 797cff7d201d968006b0b3092aa2316b554f5811