(maint) Remove redirect to PuppetDB dashboard
This commit changes the PuppetDB dashboard to be served as a file
response on "/" explicitly instead of a redirect to
"/dashboard/index.html". This allows us to refrain from passing
url-prefix to the main PuppetDB app `routes` function, as well as
modularizing the dashboard to its own handler so it can be served
separately from the PuppetDB query app.