(PDB-1269) extractable-only fields
Without this commit, our expanded fields required both the (possibly empty)
data field and either the hash or certname as an identier to form the href.
With this commit, we return essentially the desired data/href structure in the
postgres case, and in the hsqldb case, we overload the field name with the
identifier to retain the ability to create the href.