(PDB-1536) Support POST queries for nodes
Webservers have limits on the length of GET request. This commit adds to
the nodes endpoint (root of the endpoint only) the ability to instead
issue the query as a POST, with a JSON document representation of the
query as the body. This new feature needs to be added to all endpoints
eventually, but the one where we see this problem the most is the root
of the nodes endpoint, so this commit starts there. This patch puts new
code in place to centralize handling of the PuppetDB query and related
parameters so that all endpoints can share the same code for handling
GETs and POSTs of queries. Currently only the nodes endpoint calls this
new centralized code. A bigger refactor will need to be made in the
future to switch other endpoints over to the new code.