(PDB-1465) Update query-fn callback for the simplified interface
This commit updates the callback function used by the PuppetDBServer
query function to use the new simplified interface which doesn't require
a function wrapper. This commit also removes some leftover code that
wasn't picked up in a rebase for the namespace renaming, as well as
using one of the puppetdb utility functions instead of duplicating it in
the extensions code.