(PDB-1485) Add default context root `/pdb/query` for PuppetDB Query API
This commit changes PuppetDB to server its query API at `/pdb/query`
instead of `/` by default, making similar default configuration changes
to the terminus. We do not expect, nor do we want or support, users
changing this setting so this commit also removes documentation about
how to configure this from PuppetDB.
This commit also moves the PuppetDB dashboard to its own service so that
we may serve it at the route `/pdb` instead of having to move it
alongside the query API at `/pdb/query`.
This commit also serves a redirector at `/` in a web-router-service
config that will redirect requests to `/pdb/dashboard/index.html`.