(PDB-1455) Update package dependencies for Puppet 4
In Puppet 4, the package names were changed. This commit updates the
dependencies of puppetdb to accomodate this change. On debian, versioned
dependencies cannot be satisfied with a package that uses Provides tags
to enable those capabilities (this is fixed in bleeding edge deb and
ubuntu), so the debian dependencies have an `| puppet-agent` added to
cover this.