(maint) Fix sync tests
The producer_timestamp for stubbed records was being returned correctly
in the summary query, but ignored when pulling the actual record. Rework
the format of the stub-data-atom to avoid this problem..

This used to work because puppetdb core would accept all record writes
without question, but the problem was exposed when we added a proper
up-to-date check to the replace-facts command.