(maint) Fix factset up-to-date checks
To allow sync convergence when there are conflicting writes to two 
nodes, the factset up-to-date check needs to allow records whose
producer_timestamp is equal to the existing producer_timestamp. This is
the same as the replace-catalog command's current behavior. 

Also include reports in the deactivate-node up-to-date check, using
their producer_timestamp field.