(maint) don't activate the node when it's already active
Before this commit we were updating the deactivated column of certnames to null
even if it was already null. This caused unnecessary contention on the certnames
table with each command submission, causing more frequent retries than needed.

Also corrects the corresponding test logic: if nothing happens, then no records
are updated, and the result of maybe-activate-node should *not* be positive.