(maint) Remove unneccesary querying macro to improve streaming syntax
This commit removes the `with-query-results-cursor` macro, as it was
functionally unused. The macro wrapped `with-query-results-cursor*` with a
different interface, but the only thing that used the macro was
`streamed-query-results` which wrapped the macro in the original
interface of `with-query-results-cursor*`. This commit collapses these
functions/macros.
This commit also cleans up some syntax in jdbc, removing unnecessary
parens.