(maint) Remove dead code and replace testing utility functions
This commit removes dead code from each of the query/* namespaces whic
was only used in testing code. Previously we were duplicating the
efforts of the produce-streaming-body/streamed-query-results on each of
our endpoints, this commit changes the utility functions in our tests to
use the updated pattern.
This commit also changes a number of update-in calls to update.