(maint) correct race condition in core test
Before this commit we were populating master with facts, then checking that the
mirror was empty, then waiting and checking that a sync occurs. If the sync
occurred too fast, the intial emptiness check could fail. This PR changes it
so that we check that the mirror is empty before sending facts to the master.