(PDB-1455) Make start_puppetdb/install_puppetdb support legacy installation
After adapting the start_puppetdb/install_puppetdb functions to specify the
newer packages options for the module, we still need a way to flip back to
legacy support for the purposes of upgrades, at least until we ship 3.0.0.

This patch turns all the legacy support flags into 1 flag for those
functions and now also ensures that things like confdir and embedded_path
are set correctly when legacy is true.

Signed-off-by: Ken Barber <ken@bob.sh>