(PDB-1455) Install a proper ruby when using AIO on Debian
This fixes a problem with puppetlabs-concat requiring a proper ruby vm when we
use puppetlabs-postgresql. Since AIO doesn't install one by default, we install
it especially for this node combination to work around the problem.

This issue should be solved in the 2.x branch of puppetlabs-concat once
released, since puppetlabs-concat 2.x uses proper types and providers, so the
ruby code is now internally called instead of Exec'd from the agent.

Signed-off-by: Ken Barber <ken@bob.sh>